Add a lambda-in-browser.js file
This is a companion to an index.html which can <script>-import
the bundle.js that results from browserify-compiling this file.

The idea is eventually to be able to generate such an index.html
from the text in
https://gist.github.com/masak/8e082999e06bfb4d03b8d12899bbcde5
with interactive prompts in the middle where one can try out
the λ calculus.

Still some ways to go, though, since we can't evaluate lambda
expressions yet. :)